+++ admin/header.php
@@ -39,9 +39,9 @@ $current_page = getCurrentPage();
 			</a>
 			<div class="user-dropdown">
 				<span>Welcome, <?php echo $session['username']; ?></span>
-				<img class="avatar" src="<?php echo $session['avatar']; ?>" width="20" height="20">
+				<?php echo getMedia($session['avatar'], array('class'=>'avatar', 'width'=>20, 'height'=>20)); ?>
 				<ul class="user-dropdown-menu">
-					<img class="avatar-large" src="<?php echo $session['avatar']; ?>" width="100" height="100">
+					<?php echo getMedia($session['avatar'], array('class'=>'avatar-large', 'width'=>100, 'height'=>100)); ?>
 					<li><a href="profile.php">My Profile</a></li>
 					<li><a href="../login.php?action=logout">Log Out</a></li>
 				</ul>