+++ login.php
@@ -29,6 +29,7 @@ $action = $_GET['action'] ?? '';
 		<meta charset="UTF-8">
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
 		<meta name="robots" content="noindex, nofollow">
+		<meta name="theme-color" content="<?php getSetting('theme_color'); ?>">
 		<link type="image/x-icon" href="<?php echo getMediaSrc(getSetting('site_icon', false)); ?>" rel="icon">
 		<?php getStylesheet('button.min.css'); ?>
 		<?php getStylesheet('style.css'); ?>